refactor(api): add logic for ot3 calibration
The OT3 has an autocalibration routine that relies on its pipettes' (and
gripper's) capacitive sensors. To execute this routine, the
calibration_probe method is used to find the absolute position of the
deck and the absolute position of a slot machined in the deck (with a
binary search for each edge). What comes out of this routine, which is
in new free functions in a helper file, is the estimated position in
deck coordinates of the center of the calibration slot. This can then be
subtracted from the nominal position of the center of the slot to form a
pipette- and mount-specific offset that can be saved as the OT3
equivalent of pipette calibration.

This is a heavily parametrized process; everything from the initial
positions to the amount of binary searching to the exact behavior of
each capacitive sensor pass can be changed. To accomodate this, add (a
significant amount of) content to the robot config.

While adding tests for new configuration values, a little refactoring
reduced the size of the test files and added typing to some of the
config tests.

Tests for the calibration functionality itself are a little tough to
write in a robust way. find_edge has complex behavior between the error
handling and the binary search, so that is well tested; but
find_deck_position strings together a couple extremely straightforward
ot3api calls that would not be testable without writing the same code.